+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Rename "Move followers" to "Import followers".
+
 ## [2.18.0] - 2024-05-07
 
 ### Added